Move bounds tests into sources modal spec
