+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### 🐞 Bug fixes
 
+- Moved bounds field Cypress coverage into the sources modal spec.
 - Improved the local style open modal and local file upload.
 - Fixed the Expression editor (for long expressions) being able to be float under other components further down
 - Fixed an issue when clicking on a popup and then clicking on the map again
